first-version/requirements.txt
llama-parse
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langchain-huggingface==0.0.3
langchain-google-genai
faiss-cpu
sentence-transformers
rank_bm25
google-genai
pyyaml
joblib
scikit-learn
fastapi
uvicorn
python-multipart
httpx
